Move partition ID up from ResourceModifiedMessage to BaseResourceModifiedMessage.
